+++ History.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Add `saveUninitialized` option to control saving uninitialized sessions
+  * Add `unset` option to control unsetting `req.session`
   * deps: buffer-crc32@0.2.3
 
 1.3.1 / 2014-06-14